backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend server for movie click handling"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.24.0"
pydantic = "^2.4.2"
pydantic-extra-types = "^2.1.0"
email-validator = "^2.1.0.post1"
google-genai = "^1.9.0"
tenacity = "^8.2.3"
moviepy = "^1.0.3"
av = "^11.0.0"
python-dotenv = "^1.0.0"
numpy = "^1.26.0"
edge-tts = "^6.1.9"
pydub = "^0.25.1"
httpx = "*"
certifi = "*"
ffmpeg-python = "^0.2.0"
yt-dlp = "^2025.03.3"
google-cloud-storage = "^2.14.0"
boto3 = "^1.34.0"
azure-storage-blob = "^12.19.0"
minio = "^7.2.0"
aiosmtplib = "^3.0.1"
jinja2 = "^3.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 